Abstracted configs and renamed structure to viewcollective
