+++ committees/account_pages/views.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.views.generic import TemplateView
 
-from user_interaction.account_pages.mixins import AccountViewMixin
+from user_interaction.accountcollective import AccountViewMixin
 
 
 class AssociationGroupAccountView(AccountViewMixin, PermissionRequiredMixin, TemplateView):
